client: Add Delete all functionality to delete all todos with one click

+++ backend/api/controllers/todo.js
@@ -15,7 +15,20 @@ export const getTodo = async (req, res, next) => {
     });
 };
 export const deleteTodo = async (req, res) => {
-  if (req.query.id === `all`) {
+  if (req.query.id === `all` && req.query.completed === `false`) {
+    await Todo.deleteMany({ completed: false })
+      .then((ret) => {
+        console.log("Todos deleted", ret);
+        res.status(200).json("ok");
+      })
+      .catch((error) => {
+        console.log(error);
+        res.status(500).json(error);
+      });
+    return;
+  }
+
+  if (req.query.id === `all` && req.query.completed === `true`) {
     await Todo.deleteMany({ completed: true })
       .then((ret) => {
         console.log("Todos deleted", ret);
@@ -27,6 +40,7 @@ export const deleteTodo = async (req, res) => {
       });
     return;
   }
+
   await Todo.findByIdAndDelete(req.query.id)
     .then((ret) => {
       console.log("Todo deleted", ret);
